backend/requirements/base.txt
Django>=5.1,<6.0
django-cors-headers>=4.7,<5.0
psycopg[binary]>=3.2,<4.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
